Disable test_keyboardinterrupt_twice on PyPy
The test hangs, but the suite itself works on PyPy, so let us just
skip it.

+++ functional_tests/test_multiprocessing/test_keyboardinterrupt.py
@@ -83,6 +83,10 @@ def test_keyboardinterrupt():
 
 
 def test_keyboardinterrupt_twice():
+    import platform
+    if platform.python_implementation() == 'PyPy':
+        from unittest import SkipTest
+        raise SkipTest('Test hangs on PyPy')
     process, logfile, killfile = keyboardinterrupt('keyboardinterrupt_twice.py')
     waitForKillFile(killfile)
     os.killpg(process.pid, signal.SIGINT)
